feat: Moves sound playback queue to separate class.
